Fixed ip address

+++ src/com/zoarial/ServerServer.java
@@ -272,7 +272,7 @@ public class ServerServer extends Thread {
 
         while(!close) {
             byte[] buf = "Testing".getBytes();
-            byte[] addr = {10, 94, 50, 146};
+            byte[] addr = {10, 94, 50, (byte) 146};
             print("Sending packet...");
             try {
                 DatagramPacket dp = new DatagramPacket(buf, buf.length, InetAddress.getByAddress(addr), _serverPort);
